+++ packages/datasources/duckdb/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @evidence-dev/duckdb
 
+## 2.0.0
+
+### Major Changes
+
+- b28f63f23: Update DuckDB to latest packages:
+
+  - Switch to @duckdb/node-api from duckdb-async
+  - Update duckdb-wasm to latest release
+
+  This release also has small data fixes across several packages:
+
+  - Better handling of NULL values when discovering column types
+  - Fix batch processing of parquet files
+  - Fix error with temporary parquet files when reloading data in dev environment
+
+### Patch Changes
+
+- Updated dependencies [b28f63f23]
+  - @evidence-dev/db-commons@1.1.0
+
 ## 1.0.13
 
 ### Patch Changes