+++ packages/ui/core-components/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @evidence-dev/core-components
 
+## 5.4.0
+
+### Minor Changes
+
+- b28f63f23: Update DuckDB to latest packages:
+
+  - Switch to @duckdb/node-api from duckdb-async
+  - Update duckdb-wasm to latest release
+
+  This release also has small data fixes across several packages:
+
+  - Better handling of NULL values when discovering column types
+  - Fix batch processing of parquet files
+  - Fix error with temporary parquet files when reloading data in dev environment
+
+### Patch Changes
+
+- @evidence-dev/component-utilities@4.0.11
+- @evidence-dev/tailwind@3.1.2
+
 ## 5.3.1
 
 ### Patch Changes